modularize slider inputs
